Extract logic in a service + remove unused return value
