+++ FrontBundle/Resources/config/services.yml
@@ -1,6 +1,7 @@
 parameters:
     php_orchestra_front.subscriber.dynamic_routing.class: PHPOrchestra\FrontBundle\EventSubscriber\DynamicRoutingSubscriber
     php_orchestra_front.manager.dynamic_routing.class: PHPOrchestra\FrontBundle\Manager\DynamicRoutingManager
+    php_orchestra_front.manager.sitemap.class: PHPOrchestra\FrontBundle\Manager\SitemapManager
 services:
     php_orchestra_front.subscriber.dynamic_routing:
         class: %php_orchestra_front.subscriber.dynamic_routing.class%
@@ -15,3 +16,7 @@ services:
             - @php_orchestra_model.repository.site
     php_orchestra.manager.current_site:
         alias: php_orchestra_display.manager.site
+    php_orchestra_front.manager.sitemap:
+        class: %php_orchestra_front.manager.sitemap.class%
+        arguments:
+            - @php_orchestra_model.repository.node
